move DBContext to DB
