+++ server/controller/comment.go
@@ -4,10 +4,11 @@ import (
 	"net/http"
 
 	"github.com/labstack/echo/v4"
+	"github.com/teru01/image/server/database"
 	"github.com/teru01/image/server/model"
 )
 
-func CreateComment(c *model.DBContext) error {
+func CreateComment(c *database.DBContext) error {
 	var comment model.Comment
 	// json.NewDecoder(c.Request().Body).Decode(&comment)
 	if err := c.Bind(&comment); err != nil {
@@ -20,7 +21,7 @@ func CreateComment(c *model.DBContext) error {
 	return c.JSON(http.StatusCreated, data)
 }
 
-func FetchComments(c *model.DBContext) error {
+func FetchComments(c *database.DBContext) error {
 	records, err := model.FetchComments(c.Db, nil)
 	if err != nil {
 		return echo.NewHTTPError(http.StatusInternalServerError, err.Error())
