+++ server/controller/controller.go
@@ -3,13 +3,13 @@ package controller
 import (
 	"net/http"
 
-	"github.com/teru01/image/server/model"
+	"github.com/teru01/image/server/database"
 )
 
 type Err struct {
 	Description string `json: "description"`
 }
 
-func IndexGet(c *model.DBContext) error {
+func IndexGet(c *database.DBContext) error {
 	return c.String(http.StatusOK, "hello")
 }
