+++ server/form/post_form.go
@@ -1,5 +1,5 @@
 package form
 
 type PostForm struct {
-	Name     string `json:"name" form:"name" gorm:"type:varchar(255)"`
+	Name string `json:"name" form:"name" gorm:"type:varchar(255)"`
 }
